feat!: remove redundant suffix from role name
I prefer to include "Profile" in the name var, rather than have it
implicitly appended by this module.  YMMV.

BREAKING CHANGE: Any resources created by the previous version will be
destroyed and recreated, and any dependent resources will need to be
updated and possibly destroyed and recreated, as a result.

Not an issue for new users, but could be quite annoying for anyone
who has been using the module and its former behavior.  Sorry, this
is exactly why I wanted to make this change _now_, before I start using
this module everywhere.